add privacy permission dialog to SubmissionDispatcherFragment.kt (#133)
- allow non-cancelable dialogs (defaulting to true)
- move camera permission dialog to SubmissionQRCodeScanFragment.kt
- remove extra dialog for camera permission denied
