+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/ui/submission/SubmissionDispatcherFragment.kt
@@ -1,7 +1,5 @@
 package de.rki.coronawarnapp.ui.submission
 
-import android.Manifest
-import android.content.pm.PackageManager
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -9,13 +7,11 @@ import android.view.ViewGroup
 import de.rki.coronawarnapp.R
 import de.rki.coronawarnapp.databinding.FragmentSubmissionDispatcherBinding
 import de.rki.coronawarnapp.ui.BaseFragment
-import de.rki.coronawarnapp.util.CameraPermissionHelper
 import de.rki.coronawarnapp.util.DialogHelper
 
 class SubmissionDispatcherFragment : BaseFragment() {
 
     companion object {
-        private const val REQUEST_CAMERA_PERMISSION_CODE = 1
         private val TAG: String? = SubmissionDispatcherFragment::class.simpleName
     }
 
@@ -38,7 +34,7 @@ class SubmissionDispatcherFragment : BaseFragment() {
 
     private fun setButtonOnClickListener() {
         binding.submissionDispatcherQr.dispatcherCard.setOnClickListener {
-            checkForCameraPermission()
+            checkForDataPrivacyPermission()
         }
         binding.submissionDispatcherTanCode.dispatcherCard.setOnClickListener {
             doNavigate(
@@ -54,64 +50,23 @@ class SubmissionDispatcherFragment : BaseFragment() {
         }
     }
 
-    private fun checkForCameraPermission() {
-        if (!CameraPermissionHelper.hasCameraPermission(requireActivity())) {
-            if (shouldShowRequestPermissionRationale(Manifest.permission.CAMERA)) {
-                showCameraPermissionRationaleDialog()
-            } else {
-                requestPermissions(
-                    arrayOf(Manifest.permission.CAMERA),
-                    REQUEST_CAMERA_PERMISSION_CODE
-                )
-            }
-        } else {
-            cameraPermissionIsGranted()
-        }
-    }
-
-    private fun showCameraPermissionRationaleDialog() {
+    private fun checkForDataPrivacyPermission() {
         val cameraPermissionRationaleDialogInstance = DialogHelper.DialogInstance(
             requireActivity(),
-            R.string.submission_qr_code_scan_permission_rationale_dialog_headline,
-            R.string.submission_qr_code_scan_permission_rationale_dialog_body,
-            R.string.submission_qr_code_scan_permission_rationale_dialog_button_positive,
-            R.string.submission_qr_code_scan_permission_rationale_dialog_button_negative,
+            R.string.submission_dispatcher_qr_privacy_dialog_headline,
+            R.string.submission_dispatcher_qr_privacy_dialog_body,
+            R.string.submission_dispatcher_qr_privacy_dialog_button_positive,
+            R.string.submission_dispatcher_qr_privacy_dialog_button_negative,
+            true,
             {
-                requestPermissions(
-                    arrayOf(Manifest.permission.CAMERA),
-                    REQUEST_CAMERA_PERMISSION_CODE
-                )
+                privacyPermissionIsGranted()
             }
         )
 
         DialogHelper.showDialog(cameraPermissionRationaleDialogInstance)
     }
 
-    override fun onRequestPermissionsResult(
-        requestCode: Int,
-        permissions: Array<out String>,
-        grantResults: IntArray
-    ) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults)
-        if (grantResults.isNotEmpty() && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-            cameraPermissionIsGranted()
-        } else {
-            showCameraPermissionDeniedDialog()
-        }
-    }
-
-    private fun showCameraPermissionDeniedDialog() {
-        val cameraPermissionDeniedDialogInstance = DialogHelper.DialogInstance(
-            requireActivity(),
-            R.string.submission_qr_code_scan_permission_denied_dialog_headline,
-            R.string.submission_qr_code_scan_permission_denied_dialog_body,
-            R.string.submission_qr_code_scan_permission_denied_dialog_button_positive
-        )
-
-        DialogHelper.showDialog(cameraPermissionDeniedDialogInstance)
-    }
-
-    private fun cameraPermissionIsGranted() {
+    private fun privacyPermissionIsGranted() {
         doNavigate(
             SubmissionDispatcherFragmentDirections
                 .actionSubmissionDispatcherFragmentToSubmissionQRCodeScanFragment()
